move from interface to utils
