+++ src/property-filter/utils.ts
@@ -6,8 +6,21 @@ import {
   InternalFilteringOption,
   InternalFilteringProperty,
   InternalToken,
+  InternalTokenGroup,
   Token,
 } from './interfaces';
+
+export function isInternalToken(tokenOrGroup: InternalToken | InternalTokenGroup): tokenOrGroup is InternalToken {
+  const key: keyof InternalToken = 'operator';
+  return key in tokenOrGroup;
+}
+
+export function isInternalTokenGroup(
+  tokenOrGroup: InternalToken | InternalTokenGroup
+): tokenOrGroup is InternalTokenGroup {
+  const key: keyof InternalTokenGroup = 'operation';
+  return key in tokenOrGroup;
+}
 
 // Finds the longest property the filtering text starts from.
 export function matchFilteringProperty(
